Update - quicksort implemented
On running, the program loads all of the processes as objects, excludes processes with negative values (he told us to do that), and uses quicksort to sort them by arrival time. During the ticks, if there is a new process arriving, it will print the information. This is where it needs to be added to a priority queue. Then other things need to be checked, which I have general comments for, and then the stats at the bottom are printed (which I want to verify later on...).

All in all, a priority queue and the contents of the for loop need to be implemented, but that's really all that's left!
Use:
g++ -o program swill.cpp
Then:
program.exe initial_tester.txt