Okay, NOW the time quantum works?
I tried to test it and it didn't work. It should work now (I hope?)
I also found a lot of bugs with other stuff, and had to rearrange the code I had written. I won't put anything else in unless you need me to, but the things I told you worked are actually working now... I can help you merge your code with this if you have made any changes that don't merge right. Just let me know!
ALSO: This is going to be my submission for 'my due date', which is tomorrow. The code is formatted to print the stuff in the way he wanted it to look because of this.