Quick Update/Bug Note
Hello! I found a weird bug (below) and used this code to get around it when I submitted. If you can fix the bug, this version of the code has stuff that is irrelevant.
The bug:
When running on the cluster, the sdt::sort() function found in (#include <algorithms>) doesn't always sort the last thing read from the file. I don't know what causes this, or how to fix that, since it doesn't happen 50% of the time. Sometimes the sort works, sometimes it forgets the last item with no reasoning behind it or changes in the code. 

(An idea for...) The easy solution is to add a fake process at the end, which gets removed after using the sorting function. Make sure the fake process has the highest arrival number, so if the sort works it doesn't keep the fake process.

The hard solution is to figure out why this is happening in the first place, and fix it. I don't know how to do this, but it is a bug on poseidon only so it might not be worth messing with. I don't get it on my local computer, only poseidon...